zap_sock close to wait for empty sq
Modify zap_sock close to wait for empty send queue if `zap_close()` is
not called from the application callback call path. This is to ensure
that the send operations requested by the application got submitted
to the underlying socket before the socket get destroyed by close()
while not blocking the callback indefinitely. Calling `zap_close()` from
the callback path won't be blocked and entries in the send queue are not
guaranteed to be submitted before the socket is closed.

This patch has been tested with test cases in `ldms-test` repository.