Fix key collision problems of ldmsd requests
Without the patch, an LDMSD’s request context key consists of a message
number and a connection ID. The host generates the message number of
to-be-sent requests. In contrast, the peer generates the message number
of received requests. LDMSD has a single tree that contains the contexts
of both sent requests and received requests. The connection IDs in the
request sent to a peer and in the request received from the same peer
will be the same.

Moreover, the message numbers of the two requests could be the same
because the host and the peer generate them. When this happens, we would
see error messages “Duplicate message number …” or “Cannot find the
original request …”.

Also, the connection ID of an LDMS endpoint could overlap the memory
address of an ldmsd_cfg_xprt of a configuration file.

The patch adds a flag in request context keys to differentiate sent
requests, received requests, and requests from configuration files. It
prevents the key collision problem above.