Fix test_sampler `default` schema
`test_sampler` default schema did not specify the metric type in its own
metric records. As a result, the type could be any number and the
`sample()` which should increment the value of the metric failed to do
so because of the invalid type. This patch specified the metric type and
the issue went away.